@@ -1,13 +1,22 @@
+from __future__ import annotations
+
 from pathlib import Path
+from typing import TYPE_CHECKING
 
 from packaging.tags import Tag
-
 from poetry.core.packages.utils.link import Link
+
 from poetry.installation.chef import Chef
 from poetry.utils.env import MockEnv
 
 
-def test_get_cached_archive_for_link(config, mocker):
+if TYPE_CHECKING:
+    from pytest_mock import MockerFixture
+
+    from tests.conftest import Config
+
+
+def test_get_cached_archive_for_link(config: Config, mocker: MockerFixture):
     chef = Chef(
         config,
         MockEnv(
@@ -38,7 +47,7 @@ def test_get_cached_archive_for_link(config, mocker):
     assert Link("file:///foo/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl") == archive
 
 
-def test_get_cached_archives_for_link(config, mocker):
+def test_get_cached_archives_for_link(config: Config, mocker: MockerFixture):
     chef = Chef(
         config,
         MockEnv(
@@ -63,7 +72,7 @@ def test_get_cached_archives_for_link(config, mocker):
     }
 
 
-def test_get_cache_directory_for_link(config, config_cache_dir):
+def test_get_cache_directory_for_link(config: Config, config_cache_dir: Path):
     chef = Chef(
         config,
         MockEnv(
@@ -76,9 +85,8 @@ def test_get_cache_directory_for_link(config, config_cache_dir):
     )
 
     expected = Path(
-        "{}/artifacts/ba/63/13/283a3b3b7f95f05e9e6f84182d276f7bb0951d5b0cc24422b33f7a4648".format(
-            config_cache_dir.as_posix()
-        )
+        f"{config_cache_dir.as_posix()}/artifacts/ba/63/13/"
+        "283a3b3b7f95f05e9e6f84182d276f7bb0951d5b0cc24422b33f7a4648"
     )
 
-    assert expected == directory
+    assert directory == expected
