@@ -1,25 +1,35 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 import pytest
 
 from poetry.config.source import Source
 
 
+if TYPE_CHECKING:
+    from poetry.poetry import Poetry
+    from tests.types import CommandTesterFactory
+    from tests.types import ProjectFactory
+
+
 @pytest.fixture
-def source_one():
+def source_one() -> Source:
     return Source(name="one", url="https://one.com")
 
 
 @pytest.fixture
-def source_two():
+def source_two() -> Source:
     return Source(name="two", url="https://two.com")
 
 
 @pytest.fixture
-def source_default():
+def source_default() -> Source:
     return Source(name="default", url="https://default.com", default=True)
 
 
 @pytest.fixture
-def source_secondary():
+def source_secondary() -> Source:
     return Source(name="secondary", url="https://secondary.com", secondary=True)
 
 
@@ -27,7 +37,7 @@ _existing_source = Source(name="existing", url="https://existing.com")
 
 
 @pytest.fixture
-def source_existing():
+def source_existing() -> Source:
     return _existing_source
 
 
@@ -50,14 +60,17 @@ url = "{_existing_source.url}"
 
 
 @pytest.fixture
-def poetry_with_source(project_factory):
+def poetry_with_source(project_factory: ProjectFactory) -> Poetry:
     return project_factory(pyproject_content=PYPROJECT_WITH_SOURCES)
 
 
 @pytest.fixture
 def add_multiple_sources(
-    command_tester_factory, poetry_with_source, source_one, source_two
-):
+    command_tester_factory: CommandTesterFactory,
+    poetry_with_source: Poetry,
+    source_one: Source,
+    source_two: Source,
+) -> None:
     add = command_tester_factory("source add", poetry=poetry_with_source)
     for source in [source_one, source_two]:
         add.execute(f"{source.name} {source.url}")
