@@ -1,12 +1,28 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 import pytest
 
 
+if TYPE_CHECKING:
+    from cleo.testers.command_tester import CommandTester
+
+    from poetry.config.source import Source
+    from poetry.poetry import Poetry
+    from tests.types import CommandTesterFactory
+
+
 @pytest.fixture
-def tester(command_tester_factory, poetry_with_source, add_multiple_sources):
+def tester(
+    command_tester_factory: CommandTesterFactory,
+    poetry_with_source: Poetry,
+    add_multiple_sources: None,
+) -> CommandTester:
     return command_tester_factory("source show", poetry=poetry_with_source)
 
 
-def test_source_show_simple(tester):
+def test_source_show_simple(tester: CommandTester):
     tester.execute("")
 
     expected = """\
@@ -25,14 +41,13 @@ url        : https://two.com
 default    : no
 secondary  : no
 """.splitlines()
-    assert (
-        list(map(lambda l: l.strip(), tester.io.fetch_output().strip().splitlines()))
-        == expected
-    )
+    assert [
+        line.strip() for line in tester.io.fetch_output().strip().splitlines()
+    ] == expected
     assert tester.status_code == 0
 
 
-def test_source_show_one(tester, source_one):
+def test_source_show_one(tester: CommandTester, source_one: Source):
     tester.execute(f"{source_one.name}")
 
     expected = """\
@@ -41,14 +56,13 @@ url        : https://one.com
 default    : no
 secondary  : no
 """.splitlines()
-    assert (
-        list(map(lambda l: l.strip(), tester.io.fetch_output().strip().splitlines()))
-        == expected
-    )
+    assert [
+        line.strip() for line in tester.io.fetch_output().strip().splitlines()
+    ] == expected
     assert tester.status_code == 0
 
 
-def test_source_show_two(tester, source_one, source_two):
+def test_source_show_two(tester: CommandTester, source_one: Source, source_two: Source):
     tester.execute(f"{source_one.name} {source_two.name}")
 
     expected = """\
@@ -62,14 +76,13 @@ url        : https://two.com
 default    : no
 secondary  : no
 """.splitlines()
-    assert (
-        list(map(lambda l: l.strip(), tester.io.fetch_output().strip().splitlines()))
-        == expected
-    )
+    assert [
+        line.strip() for line in tester.io.fetch_output().strip().splitlines()
+    ] == expected
     assert tester.status_code == 0
 
 
-def test_source_show_error(tester):
+def test_source_show_error(tester: CommandTester):
     tester.execute("error")
     assert tester.io.fetch_error().strip() == "No source found with name(s): error"
     assert tester.status_code == 1
