@@ -1,4 +1,6 @@
-from .. import __version__
+from __future__ import annotations
+
+from tests.masonry.builders.fixtures.excluded_subpackage.example import __version__
 
 
 def test_version():
