@@ -1,11 +1,29 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 from poetry.core.packages.package import Package
+
 from poetry.factory import Factory
 from poetry.mixology.failure import SolveFailure
 from poetry.mixology.version_solver import VersionSolver
 from poetry.packages import DependencyPackage
 
 
-def add_to_repo(repository, name, version, deps=None, python=None):
+if TYPE_CHECKING:
+    from poetry.core.packages.project_package import ProjectPackage
+
+    from poetry.repositories import Repository
+    from tests.mixology.version_solver.conftest import Provider
+
+
+def add_to_repo(
+    repository: Repository,
+    name: str,
+    version: str,
+    deps: dict[str, str] | None = None,
+    python: str | None = None,
+) -> None:
     package = Package(name, version)
     if python:
         package.python_versions = python
@@ -18,13 +36,20 @@ def add_to_repo(repository, name, version, deps=None, python=None):
 
 
 def check_solver_result(
-    root, provider, result=None, error=None, tries=None, locked=None, use_latest=None
-):
+    root: ProjectPackage,
+    provider: Provider,
+    result: dict[str, str] | None = None,
+    error: str | None = None,
+    tries: int | None = None,
+    locked: dict[str, Package] | None = None,
+    use_latest: list[str] | None = None,
+) -> None:
     if locked is not None:
-        locked = {k: DependencyPackage(l.to_dependency(), l) for k, l in locked.items()}
+        locked = {
+            k: [DependencyPackage(l.to_dependency(), l)] for k, l in locked.items()
+        }
 
     solver = VersionSolver(root, provider, locked=locked, use_latest=use_latest)
-
     try:
         solution = solver.solve()
     except SolveFailure as e:
@@ -36,6 +61,11 @@ def check_solver_result(
 
             return
 
+        raise
+    except AssertionError as e:
+        if error:
+            assert str(e) == error
+            return
         raise
 
     packages = {}
