@@ -1,14 +1,20 @@
+from __future__ import annotations
+
 from pathlib import Path
+from typing import TYPE_CHECKING
+from typing import Any
 
 import pytest
 
 from cleo.io.null_io import NullIO
-
 from poetry.core.packages.dependency import Dependency
 from poetry.core.packages.package import Package
 from poetry.core.packages.project_package import ProjectPackage
+from poetry.core.packages.vcs_dependency import VCSDependency
 from poetry.core.version.markers import parse_marker
+
 from poetry.factory import Factory
+from poetry.packages import DependencyPackage
 from poetry.puzzle import Solver
 from poetry.puzzle.exceptions import SolverProblemError
 from poetry.puzzle.provider import Provider as BaseProvider
@@ -24,57 +30,85 @@ from tests.repositories.test_legacy_repository import (
 from tests.repositories.test_pypi_repository import MockRepository as MockPyPIRepository
 
 
+if TYPE_CHECKING:
+    import httpretty
+
+    from poetry.installation.operations.operation import Operation
+    from poetry.puzzle.transaction import Transaction
+
+DEFAULT_SOURCE_REF = (
+    VCSDependency("poetry", "git", "git@github.com:python-poetry/poetry.git").branch
+    or "HEAD"
+)
+
+
 class Provider(BaseProvider):
-    def set_package_python_versions(self, python_versions):
+    def set_package_python_versions(self, python_versions: str) -> None:
         self._package.python_versions = python_versions
         self._python_constraint = self._package.python_constraint
 
 
 @pytest.fixture()
-def io():
+def io() -> NullIO:
     return NullIO()
 
 
 @pytest.fixture()
-def package():
+def package() -> ProjectPackage:
     return ProjectPackage("root", "1.0")
 
 
 @pytest.fixture()
-def installed():
+def installed() -> InstalledRepository:
     return InstalledRepository()
 
 
 @pytest.fixture()
-def locked():
+def locked() -> Repository:
     return Repository()
 
 
 @pytest.fixture()
-def repo():
+def repo() -> Repository:
     return Repository()
 
 
 @pytest.fixture()
-def pool(repo):
+def pool(repo: Repository) -> Pool:
     return Pool([repo])
 
 
 @pytest.fixture()
-def solver(package, pool, installed, locked, io):
+def solver(
+    package: ProjectPackage,
+    pool: Pool,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
+) -> Solver:
     return Solver(
-        package, pool, installed, locked, io, provider=Provider(package, pool, io)
+        package,
+        pool,
+        installed,
+        locked,
+        io,
+        provider=Provider(package, pool, io, installed=installed),
     )
 
 
-def check_solver_result(ops, expected):
+def check_solver_result(
+    transaction: Transaction,
+    expected: list[dict[str, Any]],
+    synchronize: bool = False,
+) -> list[Operation]:
     for e in expected:
         if "skipped" not in e:
             e["skipped"] = False
 
     result = []
+    ops = transaction.calculate_operations(synchronize=synchronize)
     for op in ops:
-        if "update" == op.job_type:
+        if op.job_type == "update":
             result.append(
                 {
                     "job": "update",
@@ -90,31 +124,37 @@ def check_solver_result(ops, expected):
 
             result.append({"job": job, "package": op.package, "skipped": op.skipped})
 
-    assert expected == result
+    assert result == expected
+
+    return ops
 
 
-def test_solver_install_single(solver, repo, package):
+def test_solver_install_single(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
     repo.add_package(package_a)
 
-    ops = solver.solve([get_dependency("A")])
+    transaction = solver.solve([get_dependency("A")])
 
-    check_solver_result(ops, [{"job": "install", "package": package_a}])
+    check_solver_result(transaction, [{"job": "install", "package": package_a}])
 
 
-def test_solver_remove_if_no_longer_locked(solver, locked, installed):
+def test_solver_remove_if_no_longer_locked(
+    solver: Solver, locked: Repository, installed: InstalledRepository
+):
     package_a = get_package("A", "1.0")
     installed.add_package(package_a)
     locked.add_package(package_a)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(ops, [{"job": "remove", "package": package_a}])
+    check_solver_result(transaction, [{"job": "remove", "package": package_a}])
 
 
-def test_remove_non_installed(solver, repo, locked):
+def test_remove_non_installed(solver: Solver, repo: Repository, locked: Repository):
     package_a = get_package("A", "1.0")
     locked.add_package(package_a)
 
@@ -122,12 +162,14 @@ def test_remove_non_installed(solver, repo, locked):
 
     request = []
 
-    ops = solver.solve(request)
+    transaction = solver.solve(request)
 
-    check_solver_result(ops, [{"job": "remove", "package": package_a, "skipped": True}])
+    check_solver_result(transaction, [])
 
 
-def test_install_non_existing_package_fail(solver, repo, package):
+def test_install_non_existing_package_fail(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("B", "1"))
 
     package_a = get_package("A", "1.0")
@@ -137,7 +179,37 @@ def test_install_non_existing_package_fail(solver, repo, package):
         solver.solve()
 
 
-def test_solver_with_deps(solver, repo, package):
+def test_install_unpublished_package_does_not_fail(
+    installed: InstalledRepository,
+    solver: Solver,
+    repo: Repository,
+    package: ProjectPackage,
+):
+    package.add_dependency(Factory.create_dependency("B", "1"))
+
+    package_a = get_package("A", "1.0")
+    package_b = get_package("B", "1")
+    package_b.add_dependency(Factory.create_dependency("A", "1.0"))
+
+    repo.add_package(package_a)
+    installed.add_package(package_b)
+
+    transaction = solver.solve()
+
+    check_solver_result(
+        transaction,
+        [
+            {"job": "install", "package": package_a},
+            {
+                "job": "install",
+                "package": package_b,
+                "skipped": True,  # already installed
+            },
+        ],
+    )
+
+
+def test_solver_with_deps(solver: Solver, repo: Repository, package: ProjectPackage):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
@@ -148,12 +220,12 @@ def test_solver_with_deps(solver, repo, package):
     repo.add_package(package_b)
     repo.add_package(new_package_b)
 
-    package_a.requires.append(get_dependency("B", "<1.1"))
+    package_a.add_dependency(get_dependency("B", "<1.1"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_b},
             {"job": "install", "package": package_a},
@@ -161,7 +233,9 @@ def test_solver_with_deps(solver, repo, package):
     )
 
 
-def test_install_honours_not_equal(solver, repo, package):
+def test_install_honours_not_equal(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
@@ -176,12 +250,12 @@ def test_install_honours_not_equal(solver, repo, package):
     repo.add_package(new_package_b12)
     repo.add_package(new_package_b13)
 
-    package_a.requires.append(get_dependency("B", "<=1.3,!=1.3,!=1.2"))
+    package_a.add_dependency(get_dependency("B", "<=1.3,!=1.3,!=1.2"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": new_package_b11},
             {"job": "install", "package": package_a},
@@ -189,7 +263,9 @@ def test_install_honours_not_equal(solver, repo, package):
     )
 
 
-def test_install_with_deps_in_order(solver, repo, package):
+def test_install_with_deps_in_order(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(Factory.create_dependency("B", "*"))
     package.add_dependency(Factory.create_dependency("C", "*"))
@@ -201,15 +277,15 @@ def test_install_with_deps_in_order(solver, repo, package):
     repo.add_package(package_b)
     repo.add_package(package_c)
 
-    package_b.requires.append(get_dependency("A", ">=1.0"))
-    package_b.requires.append(get_dependency("C", ">=1.0"))
+    package_b.add_dependency(get_dependency("A", ">=1.0"))
+    package_b.add_dependency(get_dependency("C", ">=1.0"))
 
-    package_c.requires.append(get_dependency("A", ">=1.0"))
+    package_c.add_dependency(get_dependency("A", ">=1.0"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_a},
             {"job": "install", "package": package_c},
@@ -218,21 +294,31 @@ def test_install_with_deps_in_order(solver, repo, package):
     )
 
 
-def test_install_installed(solver, repo, installed, package):
+def test_install_installed(
+    solver: Solver,
+    repo: Repository,
+    installed: InstalledRepository,
+    package: ProjectPackage,
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
     installed.add_package(package_a)
     repo.add_package(package_a)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops, [{"job": "install", "package": package_a, "skipped": True}]
+        transaction, [{"job": "install", "package": package_a, "skipped": True}]
     )
 
 
-def test_update_installed(solver, repo, installed, package):
+def test_update_installed(
+    solver: Solver,
+    repo: Repository,
+    installed: InstalledRepository,
+    package: ProjectPackage,
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     installed.add_package(get_package("A", "1.0"))
@@ -242,14 +328,20 @@ def test_update_installed(solver, repo, installed, package):
     repo.add_package(package_a)
     repo.add_package(new_package_a)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops, [{"job": "update", "from": package_a, "to": new_package_a}]
+        transaction, [{"job": "update", "from": package_a, "to": new_package_a}]
     )
 
 
-def test_update_with_use_latest(solver, repo, installed, package, locked):
+def test_update_with_use_latest(
+    solver: Solver,
+    repo: Repository,
+    installed: InstalledRepository,
+    package: ProjectPackage,
+    locked: Repository,
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(Factory.create_dependency("B", "*"))
 
@@ -267,10 +359,10 @@ def test_update_with_use_latest(solver, repo, installed, package, locked):
     locked.add_package(package_a)
     locked.add_package(package_b)
 
-    ops = solver.solve(use_latest=[package_b.name])
+    transaction = solver.solve(use_latest=[package_b.name])
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_a, "skipped": True},
             {"job": "install", "package": new_package_b},
@@ -278,9 +370,9 @@ def test_update_with_use_latest(solver, repo, installed, package, locked):
     )
 
 
-def test_solver_sets_categories(solver, repo, package):
+def test_solver_sets_groups(solver: Solver, repo: Repository, package: ProjectPackage):
     package.add_dependency(Factory.create_dependency("A", "*"))
-    package.add_dependency(Factory.create_dependency("B", "*", category="dev"))
+    package.add_dependency(Factory.create_dependency("B", "*", groups=["dev"]))
 
     package_a = get_package("A", "1.0")
     package_b = get_package("B", "1.0")
@@ -291,10 +383,10 @@ def test_solver_sets_categories(solver, repo, package):
     repo.add_package(package_b)
     repo.add_package(package_c)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": package_c},
             {"job": "install", "package": package_a},
@@ -307,7 +399,9 @@ def test_solver_sets_categories(solver, repo, package):
     assert ops[1].package.category == "main"
 
 
-def test_solver_respects_root_package_python_versions(solver, repo, package):
+def test_solver_respects_root_package_python_versions(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     solver.provider.set_package_python_versions("~3.4")
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(Factory.create_dependency("B", "*"))
@@ -326,10 +420,10 @@ def test_solver_respects_root_package_python_versions(solver, repo, package):
     repo.add_package(package_c)
     repo.add_package(package_c11)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_c},
             {"job": "install", "package": package_a},
@@ -338,7 +432,9 @@ def test_solver_respects_root_package_python_versions(solver, repo, package):
     )
 
 
-def test_solver_fails_if_mismatch_root_python_versions(solver, repo, package):
+def test_solver_fails_if_mismatch_root_python_versions(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     solver.provider.set_package_python_versions("^3.4")
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(Factory.create_dependency("B", "*"))
@@ -358,7 +454,33 @@ def test_solver_fails_if_mismatch_root_python_versions(solver, repo, package):
         solver.solve()
 
 
-def test_solver_solves_optional_and_compatible_packages(solver, repo, package):
+def test_solver_ignores_python_restricted_if_mismatch_root_package_python_versions(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
+    solver.provider.set_package_python_versions("~3.8")
+    package.add_dependency(
+        Factory.create_dependency("A", {"version": "1.0", "python": "<3.8"})
+    )
+    package.add_dependency(
+        Factory.create_dependency(
+            "B", {"version": "1.0", "markers": "python_version < '3.8'"}
+        )
+    )
+
+    package_a = get_package("A", "1.0")
+    package_b = get_package("B", "1.0")
+
+    repo.add_package(package_a)
+    repo.add_package(package_b)
+
+    transaction = solver.solve()
+
+    check_solver_result(transaction, [])
+
+
+def test_solver_solves_optional_and_compatible_packages(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     solver.provider.set_package_python_versions("~3.4")
     package.extras["foo"] = [get_dependency("B")]
     package.add_dependency(
@@ -379,10 +501,10 @@ def test_solver_solves_optional_and_compatible_packages(solver, repo, package):
     repo.add_package(package_b)
     repo.add_package(package_c)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_c},
             {"job": "install", "package": package_a},
@@ -391,7 +513,9 @@ def test_solver_solves_optional_and_compatible_packages(solver, repo, package):
     )
 
 
-def test_solver_does_not_return_extras_if_not_requested(solver, repo, package):
+def test_solver_does_not_return_extras_if_not_requested(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(Factory.create_dependency("B", "*"))
 
@@ -405,10 +529,10 @@ def test_solver_does_not_return_extras_if_not_requested(solver, repo, package):
     repo.add_package(package_b)
     repo.add_package(package_c)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_a},
             {"job": "install", "package": package_b},
@@ -416,7 +540,9 @@ def test_solver_does_not_return_extras_if_not_requested(solver, repo, package):
     )
 
 
-def test_solver_returns_extras_if_requested(solver, repo, package):
+def test_solver_returns_extras_if_requested(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(
         Factory.create_dependency("B", {"version": "*", "extras": ["foo"]})
@@ -429,16 +555,16 @@ def test_solver_returns_extras_if_requested(solver, repo, package):
     dep = get_dependency("C", "^1.0", optional=True)
     dep.marker = parse_marker("extra == 'foo'")
     package_b.extras = {"foo": [dep]}
-    package_b.requires.append(dep)
+    package_b.add_dependency(dep)
 
     repo.add_package(package_a)
     repo.add_package(package_b)
     repo.add_package(package_c)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": package_c},
             {"job": "install", "package": package_a},
@@ -450,8 +576,13 @@ def test_solver_returns_extras_if_requested(solver, repo, package):
     assert ops[0].package.marker.is_any()
 
 
-@pytest.mark.parametrize(("enabled_extra",), [("one",), ("two",), (None,)])
-def test_solver_returns_extras_only_requested(solver, repo, package, enabled_extra):
+@pytest.mark.parametrize("enabled_extra", ["one", "two", None])
+def test_solver_returns_extras_only_requested(
+    solver: Solver,
+    repo: Repository,
+    package: ProjectPackage,
+    enabled_extra: bool | None,
+):
     extras = [enabled_extra] if enabled_extra is not None else []
 
     package.add_dependency(Factory.create_dependency("A", "*"))
@@ -474,15 +605,15 @@ def test_solver_returns_extras_only_requested(solver, repo, package, enabled_ext
 
     package_b.extras = {"one": [dep10], "two": [dep20]}
 
-    package_b.requires.append(dep10)
-    package_b.requires.append(dep20)
+    package_b.add_dependency(dep10)
+    package_b.add_dependency(dep20)
 
     repo.add_package(package_a)
     repo.add_package(package_b)
     repo.add_package(package_c10)
     repo.add_package(package_c20)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     expected = [
         {"job": "install", "package": package_a},
@@ -498,8 +629,8 @@ def test_solver_returns_extras_only_requested(solver, repo, package, enabled_ext
             },
         )
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         expected,
     )
 
@@ -507,9 +638,12 @@ def test_solver_returns_extras_only_requested(solver, repo, package, enabled_ext
     assert ops[0].package.marker.is_any()
 
 
-@pytest.mark.parametrize(("enabled_extra",), [("one",), ("two",), (None,)])
+@pytest.mark.parametrize("enabled_extra", ["one", "two", None])
 def test_solver_returns_extras_when_multiple_extras_use_same_dependency(
-    solver, repo, package, enabled_extra
+    solver: Solver,
+    repo: Repository,
+    package: ProjectPackage,
+    enabled_extra: bool | None,
 ):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
@@ -523,7 +657,7 @@ def test_solver_returns_extras_when_multiple_extras_use_same_dependency(
 
     package_b.extras = {"one": [dep], "two": [dep]}
 
-    package_b.requires.append(dep)
+    package_b.add_dependency(dep)
 
     extras = [enabled_extra] if enabled_extra is not None else []
     package_a.add_dependency(
@@ -534,7 +668,7 @@ def test_solver_returns_extras_when_multiple_extras_use_same_dependency(
     repo.add_package(package_b)
     repo.add_package(package_c)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     expected = [
         {"job": "install", "package": package_b},
@@ -544,8 +678,8 @@ def test_solver_returns_extras_when_multiple_extras_use_same_dependency(
     if enabled_extra is not None:
         expected.insert(0, {"job": "install", "package": package_c})
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         expected,
     )
 
@@ -553,9 +687,12 @@ def test_solver_returns_extras_when_multiple_extras_use_same_dependency(
     assert ops[0].package.marker.is_any()
 
 
-@pytest.mark.parametrize(("enabled_extra",), [("one",), ("two",), (None,)])
+@pytest.mark.parametrize("enabled_extra", ["one", "two", None])
 def test_solver_returns_extras_only_requested_nested(
-    solver, repo, package, enabled_extra
+    solver: Solver,
+    repo: Repository,
+    package: ProjectPackage,
+    enabled_extra: bool | None,
 ):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
@@ -574,8 +711,8 @@ def test_solver_returns_extras_only_requested_nested(
 
     package_b.extras = {"one": [dep10], "two": [dep20]}
 
-    package_b.requires.append(dep10)
-    package_b.requires.append(dep20)
+    package_b.add_dependency(dep10)
+    package_b.add_dependency(dep20)
 
     extras = [enabled_extra] if enabled_extra is not None else []
     package_a.add_dependency(
@@ -587,7 +724,7 @@ def test_solver_returns_extras_only_requested_nested(
     repo.add_package(package_c10)
     repo.add_package(package_c20)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     expected = [
         {"job": "install", "package": package_b},
@@ -603,13 +740,62 @@ def test_solver_returns_extras_only_requested_nested(
             },
         )
 
-    check_solver_result(ops, expected)
+    ops = check_solver_result(transaction, expected)
 
     assert ops[-1].package.marker.is_any()
     assert ops[0].package.marker.is_any()
 
 
-def test_solver_returns_prereleases_if_requested(solver, repo, package):
+def test_solver_finds_extras_next_to_non_extras(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
+    # Root depends on A[foo]
+    package.add_dependency(
+        Factory.create_dependency("A", {"version": "*", "extras": ["foo"]})
+    )
+
+    package_a = get_package("A", "1.0")
+    package_b = get_package("B", "1.0")
+    package_c = get_package("C", "1.0")
+    package_d = get_package("D", "1.0")
+
+    # A depends on B; A[foo] depends on B[bar].
+    package_a.add_dependency(Factory.create_dependency("B", "*"))
+    package_a.add_dependency(
+        Factory.create_dependency(
+            "B", {"version": "*", "extras": ["bar"], "markers": "extra == 'foo'"}
+        )
+    )
+    package_a.extras = {"foo": [get_dependency("B", "*")]}
+
+    # B depends on C; B[bar] depends on D.
+    package_b.add_dependency(Factory.create_dependency("C", "*"))
+    package_b.add_dependency(
+        Factory.create_dependency("D", {"version": "*", "markers": 'extra == "bar"'})
+    )
+    package_b.extras = {"bar": [get_dependency("D", "*")]}
+
+    repo.add_package(package_a)
+    repo.add_package(package_b)
+    repo.add_package(package_c)
+    repo.add_package(package_d)
+
+    transaction = solver.solve()
+
+    check_solver_result(
+        transaction,
+        [
+            {"job": "install", "package": package_c},
+            {"job": "install", "package": package_d},
+            {"job": "install", "package": package_b},
+            {"job": "install", "package": package_a},
+        ],
+    )
+
+
+def test_solver_returns_prereleases_if_requested(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(Factory.create_dependency("B", "*"))
     package.add_dependency(
@@ -626,10 +812,10 @@ def test_solver_returns_prereleases_if_requested(solver, repo, package):
     repo.add_package(package_c)
     repo.add_package(package_c_dev)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_a},
             {"job": "install", "package": package_b},
@@ -638,7 +824,9 @@ def test_solver_returns_prereleases_if_requested(solver, repo, package):
     )
 
 
-def test_solver_does_not_return_prereleases_if_not_requested(solver, repo, package):
+def test_solver_does_not_return_prereleases_if_not_requested(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(Factory.create_dependency("B", "*"))
     package.add_dependency(Factory.create_dependency("C", "*"))
@@ -653,10 +841,10 @@ def test_solver_does_not_return_prereleases_if_not_requested(solver, repo, packa
     repo.add_package(package_c)
     repo.add_package(package_c_dev)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_a},
             {"job": "install", "package": package_b},
@@ -665,7 +853,9 @@ def test_solver_does_not_return_prereleases_if_not_requested(solver, repo, packa
     )
 
 
-def test_solver_sub_dependencies_with_requirements(solver, repo, package):
+def test_solver_sub_dependencies_with_requirements(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(Factory.create_dependency("B", "*"))
 
@@ -685,10 +875,10 @@ def test_solver_sub_dependencies_with_requirements(solver, repo, package):
     repo.add_package(package_c)
     repo.add_package(package_d)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": package_d},
             {"job": "install", "package": package_c},
@@ -701,7 +891,9 @@ def test_solver_sub_dependencies_with_requirements(solver, repo, package):
     assert op.package.marker.is_any()
 
 
-def test_solver_sub_dependencies_with_requirements_complex(solver, repo, package):
+def test_solver_sub_dependencies_with_requirements_complex(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(
         Factory.create_dependency("A", {"version": "^1.0", "python": "<5.0"})
     )
@@ -743,10 +935,10 @@ def test_solver_sub_dependencies_with_requirements_complex(solver, repo, package
     repo.add_package(package_e)
     repo.add_package(package_f)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_e},
             {"job": "install", "package": package_f},
@@ -759,7 +951,7 @@ def test_solver_sub_dependencies_with_requirements_complex(solver, repo, package
 
 
 def test_solver_sub_dependencies_with_not_supported_python_version(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("^3.5")
     package.add_dependency(Factory.create_dependency("A", "*"))
@@ -775,13 +967,13 @@ def test_solver_sub_dependencies_with_not_supported_python_version(
     repo.add_package(package_a)
     repo.add_package(package_b)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(ops, [{"job": "install", "package": package_a}])
+    check_solver_result(transaction, [{"job": "install", "package": package_a}])
 
 
 def test_solver_sub_dependencies_with_not_supported_python_version_transitive(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("^3.4")
 
@@ -812,10 +1004,10 @@ def test_solver_sub_dependencies_with_not_supported_python_version_transitive(
     repo.add_package(sniffio)
     repo.add_package(sniffio_1_1_0)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": sniffio, "skipped": False},
             {"job": "install", "package": httpcore, "skipped": False},
@@ -825,13 +1017,13 @@ def test_solver_sub_dependencies_with_not_supported_python_version_transitive(
 
 
 def test_solver_with_dependency_in_both_main_and_dev_dependencies(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("^3.5")
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(
         Factory.create_dependency(
-            "A", {"version": "*", "extras": ["foo"]}, category="dev"
+            "A", {"version": "*", "extras": ["foo"]}, groups=["dev"]
         )
     )
 
@@ -854,10 +1046,10 @@ def test_solver_with_dependency_in_both_main_and_dev_dependencies(
     repo.add_package(package_c)
     repo.add_package(package_d)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": package_d},
             {"job": "install", "package": package_b},
@@ -877,14 +1069,14 @@ def test_solver_with_dependency_in_both_main_and_dev_dependencies(
     assert a.category == "main"
 
 
-def test_solver_with_dependency_in_both_main_and_dev_dependencies_with_one_more_dependent(
-    solver, repo, package
+def test_solver_with_dependency_in_both_main_and_dev_dependencies_with_one_more_dependent(  # noqa: E501
+    solver: Solver, repo: Repository, package: Package
 ):
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(Factory.create_dependency("E", "*"))
     package.add_dependency(
         Factory.create_dependency(
-            "A", {"version": "*", "extras": ["foo"]}, category="dev"
+            "A", {"version": "*", "extras": ["foo"]}, groups=["dev"]
         )
     )
 
@@ -911,10 +1103,10 @@ def test_solver_with_dependency_in_both_main_and_dev_dependencies_with_one_more_
     repo.add_package(package_d)
     repo.add_package(package_e)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": package_b},
             {"job": "install", "package": package_d},
@@ -937,7 +1129,9 @@ def test_solver_with_dependency_in_both_main_and_dev_dependencies_with_one_more_
     assert e.category == "main"
 
 
-def test_solver_with_dependency_and_prerelease_sub_dependencies(solver, repo, package):
+def test_solver_with_dependency_and_prerelease_sub_dependencies(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
@@ -951,10 +1145,10 @@ def test_solver_with_dependency_and_prerelease_sub_dependencies(solver, repo, pa
     package_b = get_package("B", "1.0.0.dev4")
     repo.add_package(package_b)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_b},
             {"job": "install", "package": package_a},
@@ -962,7 +1156,9 @@ def test_solver_with_dependency_and_prerelease_sub_dependencies(solver, repo, pa
     )
 
 
-def test_solver_circular_dependency(solver, repo, package):
+def test_solver_circular_dependency(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
@@ -978,10 +1174,10 @@ def test_solver_circular_dependency(solver, repo, package):
     repo.add_package(package_b)
     repo.add_package(package_c)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": package_c},
             {"job": "install", "package": package_b},
@@ -989,10 +1185,12 @@ def test_solver_circular_dependency(solver, repo, package):
         ],
     )
 
-    assert "main" == ops[0].package.category
+    assert ops[0].package.category == "main"
 
 
-def test_solver_circular_dependency_chain(solver, repo, package):
+def test_solver_circular_dependency_chain(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
@@ -1012,10 +1210,10 @@ def test_solver_circular_dependency_chain(solver, repo, package):
     repo.add_package(package_c)
     repo.add_package(package_d)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": package_d},
             {"job": "install", "package": package_c},
@@ -1024,10 +1222,12 @@ def test_solver_circular_dependency_chain(solver, repo, package):
         ],
     )
 
-    assert "main" == ops[0].package.category
+    assert ops[0].package.category == "main"
 
 
-def test_solver_dense_dependencies(solver, repo, package):
+def test_solver_dense_dependencies(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     # The root package depends on packages A0...An-1,
     # And package Ai depends  on packages A0...Ai-1
     # This graph is a transitive tournament
@@ -1041,14 +1241,16 @@ def test_solver_dense_dependencies(solver, repo, package):
         for j in range(i):
             package_ai.add_dependency(Factory.create_dependency("a" + str(j), "^1.0"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops, [{"job": "install", "package": packages[i]} for i in range(n)]
+        transaction, [{"job": "install", "package": packages[i]} for i in range(n)]
     )
 
 
-def test_solver_duplicate_dependencies_same_constraint(solver, repo, package):
+def test_solver_duplicate_dependencies_same_constraint(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
@@ -1064,10 +1266,10 @@ def test_solver_duplicate_dependencies_same_constraint(solver, repo, package):
     repo.add_package(package_a)
     repo.add_package(package_b)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_b},
             {"job": "install", "package": package_a},
@@ -1075,7 +1277,9 @@ def test_solver_duplicate_dependencies_same_constraint(solver, repo, package):
     )
 
 
-def test_solver_duplicate_dependencies_different_constraints(solver, repo, package):
+def test_solver_duplicate_dependencies_different_constraints(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
@@ -1093,10 +1297,10 @@ def test_solver_duplicate_dependencies_different_constraints(solver, repo, packa
     repo.add_package(package_b10)
     repo.add_package(package_b20)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_b10},
             {"job": "install", "package": package_b20},
@@ -1106,7 +1310,7 @@ def test_solver_duplicate_dependencies_different_constraints(solver, repo, packa
 
 
 def test_solver_duplicate_dependencies_different_constraints_same_requirements(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
@@ -1132,7 +1336,264 @@ So, because no versions of a match !=1.0
     assert str(e.value) == expected
 
 
-def test_solver_duplicate_dependencies_sub_dependencies(solver, repo, package):
+def test_solver_duplicate_dependencies_different_constraints_merge_by_marker(
+    solver: Solver, repo: Repository, package: Package
+):
+    package.add_dependency(Factory.create_dependency("A", "*"))
+
+    package_a = get_package("A", "1.0")
+    package_a.add_dependency(
+        Factory.create_dependency("B", {"version": "^1.0", "python": "<3.4"})
+    )
+    package_a.add_dependency(
+        Factory.create_dependency("B", {"version": "^2.0", "python": ">=3.4"})
+    )
+    package_a.add_dependency(
+        Factory.create_dependency("B", {"version": "!=1.1", "python": "<3.4"})
+    )
+
+    package_b10 = get_package("B", "1.0")
+    package_b11 = get_package("B", "1.1")
+    package_b20 = get_package("B", "2.0")
+
+    repo.add_package(package_a)
+    repo.add_package(package_b10)
+    repo.add_package(package_b11)
+    repo.add_package(package_b20)
+
+    transaction = solver.solve()
+
+    check_solver_result(
+        transaction,
+        [
+            {"job": "install", "package": package_b10},
+            {"job": "install", "package": package_b20},
+            {"job": "install", "package": package_a},
+        ],
+    )
+
+
+def test_solver_duplicate_dependencies_different_sources_types_are_preserved(
+    solver: Solver, repo: Repository, package: Package
+):
+    pendulum = get_package("pendulum", "2.0.3")
+    repo.add_package(pendulum)
+    repo.add_package(get_package("cleo", "1.0.0"))
+    repo.add_package(get_package("demo", "0.1.0"))
+
+    dependency_pypi = Factory.create_dependency("demo", ">=0.1.0")
+    dependency_git = Factory.create_dependency(
+        "demo", {"git": "https://github.com/demo/demo.git"}, groups=["dev"]
+    )
+    package.add_dependency(dependency_git)
+    package.add_dependency(dependency_pypi)
+
+    demo = Package(
+        "demo",
+        "0.1.2",
+        source_type="git",
+        source_url="https://github.com/demo/demo.git",
+        source_reference=DEFAULT_SOURCE_REF,
+        source_resolved_reference="9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
+    )
+
+    transaction = solver.solve()
+
+    ops = check_solver_result(
+        transaction,
+        [{"job": "install", "package": pendulum}, {"job": "install", "package": demo}],
+    )
+
+    op = ops[1]
+
+    assert op.package.source_type == demo.source_type
+    assert op.package.source_reference == DEFAULT_SOURCE_REF
+    assert op.package.source_resolved_reference.startswith(
+        demo.source_resolved_reference
+    )
+
+    complete_package = solver.provider.complete_package(
+        DependencyPackage(package.to_dependency(), package)
+    )
+
+    assert len(complete_package.all_requires) == 2
+
+    pypi, git = complete_package.all_requires
+
+    assert isinstance(pypi, Dependency)
+    assert pypi == dependency_pypi
+
+    assert isinstance(git, VCSDependency)
+    assert git.constraint
+    assert git.constraint != dependency_git.constraint
+    assert (git.name, git.source_type, git.source_url, git.source_reference) == (
+        dependency_git.name,
+        dependency_git.source_type,
+        dependency_git.source_url,
+        DEFAULT_SOURCE_REF,
+    )
+
+
+def test_solver_duplicate_dependencies_different_constraints_merge_no_markers(
+    solver: Solver, repo: Repository, package: Package
+):
+    package.add_dependency(Factory.create_dependency("A", "*"))
+    package.add_dependency(Factory.create_dependency("B", "1.0"))
+
+    package_a10 = get_package("A", "1.0")
+    package_a10.add_dependency(Factory.create_dependency("C", {"version": "^1.0"}))
+
+    package_a20 = get_package("A", "2.0")
+    package_a20.add_dependency(
+        Factory.create_dependency("C", {"version": "^2.0"})  # incompatible with B
+    )
+    package_a20.add_dependency(
+        Factory.create_dependency("C", {"version": "!=2.1", "python": "3.10"})
+    )
+
+    package_b = get_package("B", "1.0")
+    package_b.add_dependency(Factory.create_dependency("C", {"version": "<2.0"}))
+
+    package_c10 = get_package("C", "1.0")
+    package_c20 = get_package("C", "2.0")
+    package_c21 = get_package("C", "2.1")
+
+    repo.add_package(package_a10)
+    repo.add_package(package_a20)
+    repo.add_package(package_b)
+    repo.add_package(package_c10)
+    repo.add_package(package_c20)
+    repo.add_package(package_c21)
+
+    transaction = solver.solve()
+
+    check_solver_result(
+        transaction,
+        [
+            {"job": "install", "package": package_c10},
+            {"job": "install", "package": package_a10},  # only a10, not a20
+            {"job": "install", "package": package_b},
+        ],
+    )
+
+
+def test_solver_duplicate_dependencies_ignore_overrides_with_empty_marker_intersection(
+    solver: Solver, repo: Repository, package: Package
+):
+    """
+    Distinct requirements per marker:
+    * Python 2.7: A (which requires B) and B
+    * Python 3.6: same as Python 2.7 but with different versions
+    * Python 3.7: only A
+    * Python 3.8: only B
+    """
+    package.add_dependency(
+        Factory.create_dependency("A", {"version": "1.0", "python": "~2.7"})
+    )
+    package.add_dependency(
+        Factory.create_dependency("A", {"version": "2.0", "python": "~3.6"})
+    )
+    package.add_dependency(
+        Factory.create_dependency("A", {"version": "3.0", "python": "~3.7"})
+    )
+    package.add_dependency(
+        Factory.create_dependency("B", {"version": "1.0", "python": "~2.7"})
+    )
+    package.add_dependency(
+        Factory.create_dependency("B", {"version": "2.0", "python": "~3.6"})
+    )
+    package.add_dependency(
+        Factory.create_dependency("B", {"version": "3.0", "python": "~3.8"})
+    )
+
+    package_a10 = get_package("A", "1.0")
+    package_a10.add_dependency(
+        Factory.create_dependency("B", {"version": "^1.0", "python": "~2.7"})
+    )
+
+    package_a20 = get_package("A", "2.0")
+    package_a20.add_dependency(
+        Factory.create_dependency("B", {"version": "^2.0", "python": "~3.6"})
+    )
+
+    package_a30 = get_package("A", "3.0")  # no dep to B
+
+    package_b10 = get_package("B", "1.0")
+    package_b11 = get_package("B", "1.1")
+    package_b20 = get_package("B", "2.0")
+    package_b21 = get_package("B", "2.1")
+    package_b30 = get_package("B", "3.0")
+
+    repo.add_package(package_a10)
+    repo.add_package(package_a20)
+    repo.add_package(package_a30)
+    repo.add_package(package_b10)
+    repo.add_package(package_b11)
+    repo.add_package(package_b20)
+    repo.add_package(package_b21)
+    repo.add_package(package_b30)
+
+    transaction = solver.solve()
+
+    check_solver_result(
+        transaction,
+        [
+            {"job": "install", "package": package_b10},
+            {"job": "install", "package": package_b20},
+            {"job": "install", "package": package_a10},
+            {"job": "install", "package": package_a20},
+            {"job": "install", "package": package_a30},
+            {"job": "install", "package": package_b30},
+        ],
+    )
+
+
+def test_solver_duplicate_dependencies_ignore_overrides_with_empty_marker_intersection2(
+    solver: Solver, repo: Repository, package: Package
+):
+    """
+    Empty intersection between top level dependency and transient dependency.
+    """
+    package.add_dependency(Factory.create_dependency("A", {"version": "1.0"}))
+    package.add_dependency(
+        Factory.create_dependency("B", {"version": ">=2.0", "python": ">=3.7"})
+    )
+    package.add_dependency(
+        Factory.create_dependency("B", {"version": "*", "python": "<3.7"})
+    )
+
+    package_a10 = get_package("A", "1.0")
+    package_a10.add_dependency(
+        Factory.create_dependency("B", {"version": ">=2.0", "python": ">=3.7"})
+    )
+    package_a10.add_dependency(
+        Factory.create_dependency("B", {"version": "*", "python": "<3.7"})
+    )
+
+    package_b10 = get_package("B", "1.0")
+    package_b10.python_versions = "<3.7"
+    package_b20 = get_package("B", "2.0")
+    package_b20.python_versions = ">=3.7"
+
+    repo.add_package(package_a10)
+    repo.add_package(package_b10)
+    repo.add_package(package_b20)
+
+    transaction = solver.solve()
+
+    check_solver_result(
+        transaction,
+        [
+            {"job": "install", "package": package_b10},
+            {"job": "install", "package": package_b20},
+            {"job": "install", "package": package_a10},
+        ],
+    )
+
+
+def test_solver_duplicate_dependencies_sub_dependencies(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
@@ -1157,10 +1618,10 @@ def test_solver_duplicate_dependencies_sub_dependencies(solver, repo, package):
     repo.add_package(package_c12)
     repo.add_package(package_c15)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_c12},
             {"job": "install", "package": package_c15},
@@ -1171,7 +1632,66 @@ def test_solver_duplicate_dependencies_sub_dependencies(solver, repo, package):
     )
 
 
-def test_solver_fails_if_dependency_name_does_not_match_package(solver, repo, package):
+def test_duplicate_path_dependencies(solver: Solver, package: ProjectPackage) -> None:
+    solver.provider.set_package_python_versions("^3.7")
+    fixtures = Path(__file__).parent.parent / "fixtures"
+    project_dir = fixtures / "with_conditional_path_deps"
+
+    path1 = (project_dir / "demo_one").as_posix()
+    demo1 = Package("demo", "1.2.3", source_type="directory", source_url=path1)
+    package.add_dependency(
+        Factory.create_dependency(
+            "demo", {"path": path1, "markers": "sys_platform == 'linux'"}
+        )
+    )
+
+    path2 = (project_dir / "demo_two").as_posix()
+    demo2 = Package("demo", "1.2.3", source_type="directory", source_url=path2)
+    package.add_dependency(
+        Factory.create_dependency(
+            "demo", {"path": path2, "markers": "sys_platform == 'win32'"}
+        )
+    )
+
+    transaction = solver.solve()
+
+    check_solver_result(
+        transaction,
+        [
+            {"job": "install", "package": demo1},
+            {"job": "install", "package": demo2},
+        ],
+    )
+
+
+def test_duplicate_path_dependencies_same_path(
+    solver: Solver, package: ProjectPackage
+) -> None:
+    solver.provider.set_package_python_versions("^3.7")
+    fixtures = Path(__file__).parent.parent / "fixtures"
+    project_dir = fixtures / "with_conditional_path_deps"
+
+    path1 = (project_dir / "demo_one").as_posix()
+    demo1 = Package("demo", "1.2.3", source_type="directory", source_url=path1)
+    package.add_dependency(
+        Factory.create_dependency(
+            "demo", {"path": path1, "markers": "sys_platform == 'linux'"}
+        )
+    )
+    package.add_dependency(
+        Factory.create_dependency(
+            "demo", {"path": path1, "markers": "sys_platform == 'win32'"}
+        )
+    )
+
+    transaction = solver.solve()
+
+    check_solver_result(transaction, [{"job": "install", "package": demo1}])
+
+
+def test_solver_fails_if_dependency_name_does_not_match_package(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(
         Factory.create_dependency(
             "my-demo", {"git": "https://github.com/demo/demo.git"}
@@ -1183,7 +1703,7 @@ def test_solver_fails_if_dependency_name_does_not_match_package(solver, repo, pa
 
 
 def test_solver_does_not_get_stuck_in_recursion_on_circular_dependency(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     package_a = get_package("A", "1.0")
     package_a.add_dependency(Factory.create_dependency("B", "^1.0"))
@@ -1198,10 +1718,10 @@ def test_solver_does_not_get_stuck_in_recursion_on_circular_dependency(
 
     package.add_dependency(Factory.create_dependency("A", "^1.0"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_c},
             {"job": "install", "package": package_b},
@@ -1210,7 +1730,9 @@ def test_solver_does_not_get_stuck_in_recursion_on_circular_dependency(
     )
 
 
-def test_solver_can_resolve_git_dependencies(solver, repo, package):
+def test_solver_can_resolve_git_dependencies(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     pendulum = get_package("pendulum", "2.0.3")
     cleo = get_package("cleo", "1.0.0")
     repo.add_package(pendulum)
@@ -1220,30 +1742,32 @@ def test_solver_can_resolve_git_dependencies(solver, repo, package):
         Factory.create_dependency("demo", {"git": "https://github.com/demo/demo.git"})
     )
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     demo = Package(
         "demo",
         "0.1.2",
         source_type="git",
         source_url="https://github.com/demo/demo.git",
-        source_reference="master",
+        source_reference=DEFAULT_SOURCE_REF,
         source_resolved_reference="9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
     )
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [{"job": "install", "package": pendulum}, {"job": "install", "package": demo}],
     )
 
     op = ops[1]
 
     assert op.package.source_type == "git"
-    assert op.package.source_reference == "master"
+    assert op.package.source_reference == DEFAULT_SOURCE_REF
     assert op.package.source_resolved_reference.startswith("9cf87a2")
 
 
-def test_solver_can_resolve_git_dependencies_with_extras(solver, repo, package):
+def test_solver_can_resolve_git_dependencies_with_extras(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     pendulum = get_package("pendulum", "2.0.3")
     cleo = get_package("cleo", "1.0.0")
     repo.add_package(pendulum)
@@ -1255,19 +1779,19 @@ def test_solver_can_resolve_git_dependencies_with_extras(solver, repo, package):
         )
     )
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     demo = Package(
         "demo",
         "0.1.2",
         source_type="git",
         source_url="https://github.com/demo/demo.git",
-        source_reference="master",
+        source_reference=DEFAULT_SOURCE_REF,
         source_resolved_reference="9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
     )
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": cleo},
             {"job": "install", "package": pendulum},
@@ -1281,7 +1805,9 @@ def test_solver_can_resolve_git_dependencies_with_extras(solver, repo, package):
     [{"branch": "a-branch"}, {"tag": "a-tag"}, {"rev": "9cf8"}],
     ids=["branch", "tag", "rev"],
 )
-def test_solver_can_resolve_git_dependencies_with_ref(solver, repo, package, ref):
+def test_solver_can_resolve_git_dependencies_with_ref(
+    solver: Solver, repo: Repository, package: Package, ref: dict[str, str]
+):
     pendulum = get_package("pendulum", "2.0.3")
     cleo = get_package("cleo", "1.0.0")
     repo.add_package(pendulum)
@@ -1300,10 +1826,10 @@ def test_solver_can_resolve_git_dependencies_with_ref(solver, repo, package, ref
     git_config.update(ref)
     package.add_dependency(Factory.create_dependency("demo", git_config))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [{"job": "install", "package": pendulum}, {"job": "install", "package": demo}],
     )
 
@@ -1314,8 +1840,8 @@ def test_solver_can_resolve_git_dependencies_with_ref(solver, repo, package, ref
     assert op.package.source_resolved_reference.startswith("9cf87a2")
 
 
-def test_solver_does_not_trigger_conflict_for_python_constraint_if_python_requirement_is_compatible(
-    solver, repo, package
+def test_solver_does_not_trigger_conflict_for_python_constraint_if_python_requirement_is_compatible(  # noqa: E501
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("~2.7 || ^3.4")
     package.add_dependency(
@@ -1327,13 +1853,13 @@ def test_solver_does_not_trigger_conflict_for_python_constraint_if_python_requir
 
     repo.add_package(package_a)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(ops, [{"job": "install", "package": package_a}])
+    check_solver_result(transaction, [{"job": "install", "package": package_a}])
 
 
-def test_solver_does_not_trigger_conflict_for_python_constraint_if_python_requirement_is_compatible_multiple(
-    solver, repo, package
+def test_solver_does_not_trigger_conflict_for_python_constraint_if_python_requirement_is_compatible_multiple(  # noqa: E501
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("~2.7 || ^3.4")
     package.add_dependency(
@@ -1353,10 +1879,10 @@ def test_solver_does_not_trigger_conflict_for_python_constraint_if_python_requir
     repo.add_package(package_a)
     repo.add_package(package_b)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_b},
             {"job": "install", "package": package_a},
@@ -1364,8 +1890,8 @@ def test_solver_does_not_trigger_conflict_for_python_constraint_if_python_requir
     )
 
 
-def test_solver_triggers_conflict_for_dependency_python_not_fully_compatible_with_package_python(
-    solver, repo, package
+def test_solver_triggers_conflict_for_dependency_python_not_fully_compatible_with_package_python(  # noqa: E501
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("~2.7 || ^3.4")
     package.add_dependency(
@@ -1381,8 +1907,8 @@ def test_solver_triggers_conflict_for_dependency_python_not_fully_compatible_wit
         solver.solve()
 
 
-def test_solver_finds_compatible_package_for_dependency_python_not_fully_compatible_with_package_python(
-    solver, repo, package
+def test_solver_finds_compatible_package_for_dependency_python_not_fully_compatible_with_package_python(  # noqa: E501
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("~2.7 || ^3.4")
     package.add_dependency(
@@ -1398,13 +1924,13 @@ def test_solver_finds_compatible_package_for_dependency_python_not_fully_compati
     repo.add_package(package_a100)
     repo.add_package(package_a101)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(ops, [{"job": "install", "package": package_a100}])
+    check_solver_result(transaction, [{"job": "install", "package": package_a100}])
 
 
-def test_solver_does_not_trigger_new_resolution_on_duplicate_dependencies_if_only_extras(
-    solver, repo, package
+def test_solver_does_not_trigger_new_resolution_on_duplicate_dependencies_if_only_extras(  # noqa: E501
+    solver: Solver, repo: Repository, package: Package
 ):
     dep1 = Dependency.create_from_pep_508('B (>=1.0); extra == "foo"')
     dep1.activate()
@@ -1417,8 +1943,8 @@ def test_solver_does_not_trigger_new_resolution_on_duplicate_dependencies_if_onl
 
     package_a = get_package("A", "1.0.0")
     package_a.extras = {"foo": [dep1], "bar": [dep2]}
-    package_a.requires.append(dep1)
-    package_a.requires.append(dep2)
+    package_a.add_dependency(dep1)
+    package_a.add_dependency(dep2)
 
     package_b2 = get_package("B", "2.0.0")
     package_b1 = get_package("B", "1.0.0")
@@ -1427,10 +1953,10 @@ def test_solver_does_not_trigger_new_resolution_on_duplicate_dependencies_if_onl
     repo.add_package(package_b1)
     repo.add_package(package_b2)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": package_b2},
             {"job": "install", "package": package_a},
@@ -1442,7 +1968,7 @@ def test_solver_does_not_trigger_new_resolution_on_duplicate_dependencies_if_onl
 
 
 def test_solver_does_not_raise_conflict_for_locked_conditional_dependencies(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("~2.7 || ^3.4")
     package.add_dependency(
@@ -1462,10 +1988,10 @@ def test_solver_does_not_raise_conflict_for_locked_conditional_dependencies(
     repo.add_package(package_b)
 
     solver._locked = Repository([package_a])
-    ops = solver.solve(use_latest=[package_b.name])
+    transaction = solver.solve(use_latest=[package_b.name])
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_a},
             {"job": "install", "package": package_b},
@@ -1474,7 +2000,7 @@ def test_solver_does_not_raise_conflict_for_locked_conditional_dependencies(
 
 
 def test_solver_returns_extras_if_requested_in_dependencies_and_not_in_root_package(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     package.add_dependency(Factory.create_dependency("A", "*"))
     package.add_dependency(Factory.create_dependency("B", "*"))
@@ -1499,10 +2025,10 @@ def test_solver_returns_extras_if_requested_in_dependencies_and_not_in_root_pack
     repo.add_package(package_c)
     repo.add_package(package_d)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_d},
             {"job": "install", "package": package_c},
@@ -1513,7 +2039,7 @@ def test_solver_returns_extras_if_requested_in_dependencies_and_not_in_root_pack
 
 
 def test_solver_should_not_resolve_prerelease_version_if_not_requested(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     package.add_dependency(Factory.create_dependency("A", "~1.8.0"))
     package.add_dependency(Factory.create_dependency("B", "^0.5.0"))
@@ -1532,16 +2058,17 @@ def test_solver_should_not_resolve_prerelease_version_if_not_requested(
 
 
 def test_solver_ignores_dependencies_with_incompatible_python_full_version_marker(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("^3.6")
     package.add_dependency(Factory.create_dependency("A", "^1.0"))
     package.add_dependency(Factory.create_dependency("B", "^2.0"))
 
     package_a = get_package("A", "1.0.0")
-    package_a.requires.append(
+    package_a.add_dependency(
         Dependency.create_from_pep_508(
-            'B (<2.0); platform_python_implementation == "PyPy" and python_full_version < "2.7.9"'
+            'B (<2.0); platform_python_implementation == "PyPy" and python_full_version'
+            ' < "2.7.9"'
         )
     )
 
@@ -1552,10 +2079,10 @@ def test_solver_ignores_dependencies_with_incompatible_python_full_version_marke
     repo.add_package(package_b100)
     repo.add_package(package_b200)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_a},
             {"job": "install", "package": package_b200},
@@ -1563,7 +2090,9 @@ def test_solver_ignores_dependencies_with_incompatible_python_full_version_marke
     )
 
 
-def test_solver_git_dependencies_update(solver, repo, package, installed):
+def test_solver_git_dependencies_update(
+    solver: Solver, repo: Repository, package: Package, installed: InstalledRepository
+):
     pendulum = get_package("pendulum", "2.0.3")
     cleo = get_package("cleo", "1.0.0")
     repo.add_package(pendulum)
@@ -1574,7 +2103,7 @@ def test_solver_git_dependencies_update(solver, repo, package, installed):
         "0.1.2",
         source_type="git",
         source_url="https://github.com/demo/demo.git",
-        source_reference="master",
+        source_reference=DEFAULT_SOURCE_REF,
         source_resolved_reference="123456",
     )
     demo = Package(
@@ -1582,7 +2111,7 @@ def test_solver_git_dependencies_update(solver, repo, package, installed):
         "0.1.2",
         source_type="git",
         source_url="https://github.com/demo/demo.git",
-        source_reference="master",
+        source_reference=DEFAULT_SOURCE_REF,
         source_resolved_reference="9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
     )
     installed.add_package(demo_installed)
@@ -1591,10 +2120,10 @@ def test_solver_git_dependencies_update(solver, repo, package, installed):
         Factory.create_dependency("demo", {"git": "https://github.com/demo/demo.git"})
     )
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": pendulum},
             {"job": "update", "from": demo_installed, "to": demo},
@@ -1605,44 +2134,13 @@ def test_solver_git_dependencies_update(solver, repo, package, installed):
 
     assert op.job_type == "update"
     assert op.package.source_type == "git"
-    assert op.package.source_reference == "master"
+    assert op.package.source_reference == DEFAULT_SOURCE_REF
     assert op.package.source_resolved_reference.startswith("9cf87a2")
     assert op.initial_package.source_resolved_reference == "123456"
 
 
-def test_solver_git_dependencies_update_skipped(solver, repo, package, installed):
-    pendulum = get_package("pendulum", "2.0.3")
-    cleo = get_package("cleo", "1.0.0")
-    repo.add_package(pendulum)
-    repo.add_package(cleo)
-
-    demo = Package(
-        "demo",
-        "0.1.2",
-        source_type="git",
-        source_url="https://github.com/demo/demo.git",
-        source_reference="master",
-        source_resolved_reference="9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
-    )
-    installed.add_package(demo)
-
-    package.add_dependency(
-        Factory.create_dependency("demo", {"git": "https://github.com/demo/demo.git"})
-    )
-
-    ops = solver.solve()
-
-    check_solver_result(
-        ops,
-        [
-            {"job": "install", "package": pendulum},
-            {"job": "install", "package": demo, "skipped": True},
-        ],
-    )
-
-
-def test_solver_git_dependencies_short_hash_update_skipped(
-    solver, repo, package, installed
+def test_solver_git_dependencies_update_skipped(
+    solver: Solver, repo: Repository, package: Package, installed: InstalledRepository
 ):
     pendulum = get_package("pendulum", "2.0.3")
     cleo = get_package("cleo", "1.0.0")
@@ -1660,15 +2158,48 @@ def test_solver_git_dependencies_short_hash_update_skipped(
     installed.add_package(demo)
 
     package.add_dependency(
+        Factory.create_dependency("demo", {"git": "https://github.com/demo/demo.git"})
+    )
+
+    transaction = solver.solve()
+
+    check_solver_result(
+        transaction,
+        [
+            {"job": "install", "package": pendulum},
+            {"job": "install", "package": demo, "skipped": True},
+        ],
+    )
+
+
+def test_solver_git_dependencies_short_hash_update_skipped(
+    solver: Solver, repo: Repository, package: Package, installed: InstalledRepository
+):
+    pendulum = get_package("pendulum", "2.0.3")
+    cleo = get_package("cleo", "1.0.0")
+    repo.add_package(pendulum)
+    repo.add_package(cleo)
+
+    demo = Package(
+        "demo",
+        "0.1.2",
+        source_type="git",
+        source_url="https://github.com/demo/demo.git",
+        source_reference="9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
+        source_resolved_reference="9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
+    )
+    installed.add_package(demo)
+
+    package.add_dependency(
         Factory.create_dependency(
             "demo", {"git": "https://github.com/demo/demo.git", "rev": "9cf87a2"}
         )
     )
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": pendulum},
             {
@@ -1679,7 +2210,9 @@ def test_solver_git_dependencies_short_hash_update_skipped(
                     source_type="git",
                     source_url="https://github.com/demo/demo.git",
                     source_reference="9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
-                    source_resolved_reference="9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
+                    source_resolved_reference=(
+                        "9cf87a285a2d3fbb0b9fa621997b3acc3631ed24"
+                    ),
                 ),
                 "skipped": True,
             },
@@ -1687,7 +2220,9 @@ def test_solver_git_dependencies_short_hash_update_skipped(
     )
 
 
-def test_solver_can_resolve_directory_dependencies(solver, repo, package):
+def test_solver_can_resolve_directory_dependencies(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     pendulum = get_package("pendulum", "2.0.3")
     repo.add_package(pendulum)
 
@@ -1702,12 +2237,12 @@ def test_solver_can_resolve_directory_dependencies(solver, repo, package):
 
     package.add_dependency(Factory.create_dependency("demo", {"path": path}))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     demo = Package("demo", "0.1.2", source_type="directory", source_url=path)
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [{"job": "install", "package": pendulum}, {"job": "install", "package": demo}],
     )
 
@@ -1720,7 +2255,11 @@ def test_solver_can_resolve_directory_dependencies(solver, repo, package):
 
 
 def test_solver_can_resolve_directory_dependencies_nested_editable(
-    solver, repo, pool, installed, locked, io
+    repo: Repository,
+    pool: Pool,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
 ):
     base = Path(__file__).parent.parent / "fixtures" / "project_with_nested_local"
     poetry = Factory().create_poetry(cwd=base)
@@ -1730,10 +2269,10 @@ def test_solver_can_resolve_directory_dependencies_nested_editable(
         package, pool, installed, locked, io, provider=Provider(package, pool, io)
     )
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {
                 "job": "install",
@@ -1773,7 +2312,9 @@ def test_solver_can_resolve_directory_dependencies_nested_editable(
         assert op.package.develop is True
 
 
-def test_solver_can_resolve_directory_dependencies_with_extras(solver, repo, package):
+def test_solver_can_resolve_directory_dependencies_with_extras(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     pendulum = get_package("pendulum", "2.0.3")
     cleo = get_package("cleo", "1.0.0")
     repo.add_package(pendulum)
@@ -1792,12 +2333,12 @@ def test_solver_can_resolve_directory_dependencies_with_extras(solver, repo, pac
         Factory.create_dependency("demo", {"path": path, "extras": ["foo"]})
     )
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     demo = Package("demo", "0.1.2", source_type="directory", source_url=path)
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": cleo},
             {"job": "install", "package": pendulum},
@@ -1813,7 +2354,9 @@ def test_solver_can_resolve_directory_dependencies_with_extras(solver, repo, pac
     assert op.package.source_url == path
 
 
-def test_solver_can_resolve_sdist_dependencies(solver, repo, package):
+def test_solver_can_resolve_sdist_dependencies(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     pendulum = get_package("pendulum", "2.0.3")
     repo.add_package(pendulum)
 
@@ -1826,12 +2369,12 @@ def test_solver_can_resolve_sdist_dependencies(solver, repo, package):
 
     package.add_dependency(Factory.create_dependency("demo", {"path": path}))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     demo = Package("demo", "0.1.0", source_type="file", source_url=path)
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [{"job": "install", "package": pendulum}, {"job": "install", "package": demo}],
     )
 
@@ -1843,7 +2386,9 @@ def test_solver_can_resolve_sdist_dependencies(solver, repo, package):
     assert op.package.source_url == path
 
 
-def test_solver_can_resolve_sdist_dependencies_with_extras(solver, repo, package):
+def test_solver_can_resolve_sdist_dependencies_with_extras(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     pendulum = get_package("pendulum", "2.0.3")
     cleo = get_package("cleo", "1.0.0")
     repo.add_package(pendulum)
@@ -1860,12 +2405,12 @@ def test_solver_can_resolve_sdist_dependencies_with_extras(solver, repo, package
         Factory.create_dependency("demo", {"path": path, "extras": ["foo"]})
     )
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     demo = Package("demo", "0.1.0", source_type="file", source_url=path)
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": cleo},
             {"job": "install", "package": pendulum},
@@ -1881,7 +2426,9 @@ def test_solver_can_resolve_sdist_dependencies_with_extras(solver, repo, package
     assert op.package.source_url == path
 
 
-def test_solver_can_resolve_wheel_dependencies(solver, repo, package):
+def test_solver_can_resolve_wheel_dependencies(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     pendulum = get_package("pendulum", "2.0.3")
     repo.add_package(pendulum)
 
@@ -1894,12 +2441,12 @@ def test_solver_can_resolve_wheel_dependencies(solver, repo, package):
 
     package.add_dependency(Factory.create_dependency("demo", {"path": path}))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     demo = Package("demo", "0.1.0", source_type="file", source_url=path)
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [{"job": "install", "package": pendulum}, {"job": "install", "package": demo}],
     )
 
@@ -1911,7 +2458,9 @@ def test_solver_can_resolve_wheel_dependencies(solver, repo, package):
     assert op.package.source_url == path
 
 
-def test_solver_can_resolve_wheel_dependencies_with_extras(solver, repo, package):
+def test_solver_can_resolve_wheel_dependencies_with_extras(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     pendulum = get_package("pendulum", "2.0.3")
     cleo = get_package("cleo", "1.0.0")
     repo.add_package(pendulum)
@@ -1928,12 +2477,12 @@ def test_solver_can_resolve_wheel_dependencies_with_extras(solver, repo, package
         Factory.create_dependency("demo", {"path": path, "extras": ["foo"]})
     )
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     demo = Package("demo", "0.1.0", source_type="file", source_url=path)
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {"job": "install", "package": cleo},
             {"job": "install", "package": pendulum},
@@ -1950,7 +2499,10 @@ def test_solver_can_resolve_wheel_dependencies_with_extras(solver, repo, package
 
 
 def test_solver_can_solve_with_legacy_repository_using_proper_dists(
-    package, installed, locked, io
+    package: ProjectPackage,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
 ):
     repo = MockLegacyRepository()
     pool = Pool([repo])
@@ -1959,10 +2511,10 @@ def test_solver_can_solve_with_legacy_repository_using_proper_dists(
 
     package.add_dependency(Factory.create_dependency("isort", "4.3.4"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {
                 "job": "install",
@@ -1992,7 +2544,10 @@ def test_solver_can_solve_with_legacy_repository_using_proper_dists(
 
 
 def test_solver_can_solve_with_legacy_repository_using_proper_python_compatible_dists(
-    package, installed, locked, io
+    package: ProjectPackage,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
 ):
     package.python_versions = "^3.7"
 
@@ -2003,10 +2558,10 @@ def test_solver_can_solve_with_legacy_repository_using_proper_python_compatible_
 
     package.add_dependency(Factory.create_dependency("isort", "4.3.4"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {
                 "job": "install",
@@ -2022,7 +2577,12 @@ def test_solver_can_solve_with_legacy_repository_using_proper_python_compatible_
     )
 
 
-def test_solver_skips_invalid_versions(package, installed, locked, io):
+def test_solver_skips_invalid_versions(
+    package: ProjectPackage,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
+):
     package.python_versions = "^3.7"
 
     repo = MockPyPIRepository()
@@ -2032,14 +2592,16 @@ def test_solver_skips_invalid_versions(package, installed, locked, io):
 
     package.add_dependency(Factory.create_dependency("trackpy", "^0.4"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops, [{"job": "install", "package": get_package("trackpy", "0.4.1")}]
+        transaction, [{"job": "install", "package": get_package("trackpy", "0.4.1")}]
     )
 
 
-def test_multiple_constraints_on_root(package, solver, repo):
+def test_multiple_constraints_on_root(
+    package: ProjectPackage, solver: Solver, repo: Repository
+):
     package.add_dependency(
         Factory.create_dependency("foo", {"version": "^1.0", "python": "^2.7"})
     )
@@ -2053,16 +2615,19 @@ def test_multiple_constraints_on_root(package, solver, repo):
     repo.add_package(foo15)
     repo.add_package(foo25)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [{"job": "install", "package": foo15}, {"job": "install", "package": foo25}],
     )
 
 
 def test_solver_chooses_most_recent_version_amongst_repositories(
-    package, installed, locked, io
+    package: ProjectPackage,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
 ):
     package.python_versions = "^3.7"
     package.add_dependency(Factory.create_dependency("tomlkit", {"version": "^0.5"}))
@@ -2072,10 +2637,10 @@ def test_solver_chooses_most_recent_version_amongst_repositories(
 
     solver = Solver(package, pool, installed, locked, io)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops, [{"job": "install", "package": get_package("tomlkit", "0.5.3")}]
+    ops = check_solver_result(
+        transaction, [{"job": "install", "package": get_package("tomlkit", "0.5.3")}]
     )
 
     assert ops[0].package.source_type is None
@@ -2083,7 +2648,10 @@ def test_solver_chooses_most_recent_version_amongst_repositories(
 
 
 def test_solver_chooses_from_correct_repository_if_forced(
-    package, installed, locked, io
+    package: ProjectPackage,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
 ):
     package.python_versions = "^3.7"
     package.add_dependency(
@@ -2095,10 +2663,10 @@ def test_solver_chooses_from_correct_repository_if_forced(
 
     solver = Solver(package, pool, installed, locked, io)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {
                 "job": "install",
@@ -2113,11 +2681,14 @@ def test_solver_chooses_from_correct_repository_if_forced(
         ],
     )
 
-    assert "http://legacy.foo.bar" == ops[0].package.source_url
+    assert ops[0].package.source_url == "http://legacy.foo.bar"
 
 
 def test_solver_chooses_from_correct_repository_if_forced_and_transitive_dependency(
-    package, installed, locked, io
+    package: ProjectPackage,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
 ):
     package.python_versions = "^3.7"
     package.add_dependency(Factory.create_dependency("foo", "^1.0"))
@@ -2133,10 +2704,10 @@ def test_solver_chooses_from_correct_repository_if_forced_and_transitive_depende
 
     solver = Solver(package, pool, installed, locked, io)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {
                 "job": "install",
@@ -2152,14 +2723,17 @@ def test_solver_chooses_from_correct_repository_if_forced_and_transitive_depende
         ],
     )
 
-    assert "http://legacy.foo.bar" == ops[0].package.source_url
+    assert ops[0].package.source_url == "http://legacy.foo.bar"
 
     assert ops[1].package.source_type is None
     assert ops[1].package.source_url is None
 
 
 def test_solver_does_not_choose_from_secondary_repository_by_default(
-    package, installed, locked, io
+    package: ProjectPackage,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
 ):
     package.python_versions = "^3.7"
     package.add_dependency(Factory.create_dependency("clikit", {"version": "^0.2.0"}))
@@ -2170,10 +2744,10 @@ def test_solver_does_not_choose_from_secondary_repository_by_default(
 
     solver = Solver(package, pool, installed, locked, io)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {
                 "job": "install",
@@ -2199,13 +2773,18 @@ def test_solver_does_not_choose_from_secondary_repository_by_default(
         ],
     )
 
-    assert "http://legacy.foo.bar" == ops[0].package.source_url
+    assert ops[0].package.source_url == "http://legacy.foo.bar"
     assert ops[1].package.source_type is None
     assert ops[1].package.source_url is None
-    assert "http://legacy.foo.bar" == ops[2].package.source_url
+    assert ops[2].package.source_url == "http://legacy.foo.bar"
 
 
-def test_solver_chooses_from_secondary_if_explicit(package, installed, locked, io):
+def test_solver_chooses_from_secondary_if_explicit(
+    package: ProjectPackage,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
+):
     package.python_versions = "^3.7"
     package.add_dependency(
         Factory.create_dependency("clikit", {"version": "^0.2.0", "source": "PyPI"})
@@ -2217,10 +2796,10 @@ def test_solver_chooses_from_secondary_if_explicit(package, installed, locked, i
 
     solver = Solver(package, pool, installed, locked, io)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(
-        ops,
+    ops = check_solver_result(
+        transaction,
         [
             {
                 "job": "install",
@@ -2237,7 +2816,7 @@ def test_solver_chooses_from_secondary_if_explicit(package, installed, locked, i
         ],
     )
 
-    assert "http://legacy.foo.bar" == ops[0].package.source_url
+    assert ops[0].package.source_url == "http://legacy.foo.bar"
     assert ops[1].package.source_type is None
     assert ops[1].package.source_url is None
     assert ops[2].package.source_type is None
@@ -2245,7 +2824,12 @@ def test_solver_chooses_from_secondary_if_explicit(package, installed, locked, i
 
 
 def test_solver_discards_packages_with_empty_markers(
-    package, installed, locked, io, pool, repo
+    package: ProjectPackage,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
+    pool: Pool,
+    repo: Repository,
 ):
     package.python_versions = "~2.7 || ^3.4"
     package.add_dependency(
@@ -2269,10 +2853,10 @@ def test_solver_discards_packages_with_empty_markers(
 
     solver = Solver(package, pool, installed, locked, io)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_c},
             {"job": "install", "package": package_a},
@@ -2281,17 +2865,17 @@ def test_solver_discards_packages_with_empty_markers(
 
 
 def test_solver_does_not_raise_conflict_for_conditional_dev_dependencies(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("~2.7 || ^3.5")
     package.add_dependency(
         Factory.create_dependency(
-            "A", {"version": "^1.0", "python": "~2.7"}, category="dev"
+            "A", {"version": "^1.0", "python": "~2.7"}, groups=["dev"]
         )
     )
     package.add_dependency(
         Factory.create_dependency(
-            "A", {"version": "^2.0", "python": "^3.5"}, category="dev"
+            "A", {"version": "^2.0", "python": "^3.5"}, groups=["dev"]
         )
     )
 
@@ -2301,10 +2885,10 @@ def test_solver_does_not_raise_conflict_for_conditional_dev_dependencies(
     repo.add_package(package_a100)
     repo.add_package(package_a200)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_a100},
             {"job": "install", "package": package_a200},
@@ -2313,7 +2897,7 @@ def test_solver_does_not_raise_conflict_for_conditional_dev_dependencies(
 
 
 def test_solver_does_not_loop_indefinitely_on_duplicate_constraints_with_extras(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("~2.7 || ^3.5")
     package.add_dependency(
@@ -2335,16 +2919,21 @@ def test_solver_does_not_loop_indefinitely_on_duplicate_constraints_with_extras(
     repo.add_package(requests)
     repo.add_package(idna)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [{"job": "install", "package": idna}, {"job": "install", "package": requests}],
     )
 
 
 def test_solver_does_not_fail_with_locked_git_and_non_git_dependencies(
-    solver, repo, package, locked, pool, installed, io
+    repo: Repository,
+    package: Package,
+    locked: Repository,
+    pool: Pool,
+    installed: InstalledRepository,
+    io: NullIO,
 ):
     package.add_dependency(
         Factory.create_dependency("demo", {"git": "https://github.com/demo/demo.git"})
@@ -2356,8 +2945,8 @@ def test_solver_does_not_fail_with_locked_git_and_non_git_dependencies(
         "0.1.2",
         source_type="git",
         source_url="https://github.com/demo/demo.git",
-        source_reference="master",
-        source_resolved_reference="commit",
+        source_reference=DEFAULT_SOURCE_REF,
+        source_resolved_reference="9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
     )
 
     installed.add_package(git_package)
@@ -2370,10 +2959,10 @@ def test_solver_does_not_fail_with_locked_git_and_non_git_dependencies(
 
     solver = Solver(package, pool, installed, locked, io)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": get_package("a", "1.2.3")},
             {"job": "install", "package": git_package, "skipped": True},
@@ -2381,7 +2970,9 @@ def test_solver_does_not_fail_with_locked_git_and_non_git_dependencies(
     )
 
 
-def test_ignore_python_constraint_no_overlap_dependencies(solver, repo, package):
+def test_ignore_python_constraint_no_overlap_dependencies(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     pytest = get_package("demo", "1.0.0")
     pytest.add_dependency(
         Factory.create_dependency(
@@ -2396,16 +2987,16 @@ def test_ignore_python_constraint_no_overlap_dependencies(solver, repo, package)
     repo.add_package(pytest)
     repo.add_package(get_package("configparser", "1.2.3"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [{"job": "install", "package": pytest}],
     )
 
 
 def test_solver_should_not_go_into_an_infinite_loop_on_duplicate_dependencies(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     solver.provider.set_package_python_versions("~2.7 || ^3.5")
     package.add_dependency(Factory.create_dependency("A", "^1.0"))
@@ -2425,10 +3016,10 @@ def test_solver_should_not_go_into_an_infinite_loop_on_duplicate_dependencies(
     repo.add_package(package_b10)
     repo.add_package(package_b20)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_b10},
             {"job": "install", "package": package_b20},
@@ -2437,31 +3028,43 @@ def test_solver_should_not_go_into_an_infinite_loop_on_duplicate_dependencies(
     )
 
 
-def test_solver_remove_untracked_single(package, pool, installed, locked, io):
-    solver = Solver(package, pool, installed, locked, io, remove_untracked=True)
+def test_solver_synchronize_single(
+    package: ProjectPackage,
+    pool: Pool,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
+):
+    solver = Solver(package, pool, installed, locked, io)
     package_a = get_package("a", "1.0")
     installed.add_package(package_a)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(ops, [{"job": "remove", "package": package_a}])
+    check_solver_result(
+        transaction, [{"job": "remove", "package": package_a}], synchronize=True
+    )
 
 
 @pytest.mark.skip(reason="Poetry no longer has critical package requirements")
-def test_solver_remove_untracked_keeps_critical_package(
-    package, pool, installed, locked, io
+def test_solver_with_synchronization_keeps_critical_package(
+    package: ProjectPackage,
+    pool: Pool,
+    installed: InstalledRepository,
+    locked: Repository,
+    io: NullIO,
 ):
-    solver = Solver(package, pool, installed, locked, io, remove_untracked=True)
+    solver = Solver(package, pool, installed, locked, io)
     package_pip = get_package("setuptools", "1.0")
     installed.add_package(package_pip)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(ops, [])
+    check_solver_result(transaction, [])
 
 
 def test_solver_cannot_choose_another_version_for_directory_dependencies(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     pendulum = get_package("pendulum", "2.0.3")
     demo = get_package("demo", "0.1.0")
@@ -2490,7 +3093,7 @@ def test_solver_cannot_choose_another_version_for_directory_dependencies(
 
 
 def test_solver_cannot_choose_another_version_for_file_dependencies(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     pendulum = get_package("pendulum", "2.0.3")
     demo = get_package("demo", "0.0.8")
@@ -2517,7 +3120,7 @@ def test_solver_cannot_choose_another_version_for_file_dependencies(
 
 
 def test_solver_cannot_choose_another_version_for_git_dependencies(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     pendulum = get_package("pendulum", "2.0.3")
     demo = get_package("demo", "0.0.8")
@@ -2539,7 +3142,10 @@ def test_solver_cannot_choose_another_version_for_git_dependencies(
 
 
 def test_solver_cannot_choose_another_version_for_url_dependencies(
-    solver, repo, package, http
+    solver: Solver,
+    repo: Repository,
+    package: Package,
+    http: type[httpretty.httpretty],
 ):
     path = (
         Path(__file__).parent.parent
@@ -2577,7 +3183,7 @@ def test_solver_cannot_choose_another_version_for_url_dependencies(
 
 
 def test_solver_should_not_update_same_version_packages_if_installed_has_no_source_type(
-    solver, repo, package, installed
+    solver: Solver, repo: Repository, package: Package, installed: InstalledRepository
 ):
     package.add_dependency(Factory.create_dependency("foo", "1.0.0"))
 
@@ -2591,13 +3197,15 @@ def test_solver_should_not_update_same_version_packages_if_installed_has_no_sour
     repo.add_package(foo)
     installed.add_package(get_package("foo", "1.0.0"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(ops, [{"job": "install", "package": foo, "skipped": True}])
+    check_solver_result(
+        transaction, [{"job": "install", "package": foo, "skipped": True}]
+    )
 
 
 def test_solver_should_use_the_python_constraint_from_the_environment_if_available(
-    solver, repo, package, installed
+    solver: Solver, repo: Repository, package: Package, installed: InstalledRepository
 ):
     solver.provider.set_package_python_versions("~2.7 || ^3.5")
     package.add_dependency(Factory.create_dependency("A", "^1.0"))
@@ -2615,16 +3223,16 @@ def test_solver_should_use_the_python_constraint_from_the_environment_if_availab
     repo.add_package(b)
 
     with solver.use_environment(MockEnv((2, 7, 18))):
-        ops = solver.solve()
+        transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [{"job": "install", "package": b}, {"job": "install", "package": a}],
     )
 
 
 def test_solver_should_resolve_all_versions_for_multiple_duplicate_dependencies(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     package.python_versions = "~2.7 || ^3.5"
     package.add_dependency(
@@ -2658,10 +3266,10 @@ def test_solver_should_resolve_all_versions_for_multiple_duplicate_dependencies(
     repo.add_package(package_b30)
     repo.add_package(package_b40)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": package_a10},
             {"job": "install", "package": package_a20},
@@ -2672,7 +3280,7 @@ def test_solver_should_resolve_all_versions_for_multiple_duplicate_dependencies(
 
 
 def test_solver_should_not_raise_errors_for_irrelevant_python_constraints(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     package.python_versions = "^3.6"
     solver.provider.set_package_python_versions("^3.6")
@@ -2684,12 +3292,14 @@ def test_solver_should_not_raise_errors_for_irrelevant_python_constraints(
     dataclasses.python_versions = ">=3.6, <3.7"
 
     repo.add_package(dataclasses)
-    ops = solver.solve()
+    transaction = solver.solve()
 
-    check_solver_result(ops, [{"job": "install", "package": dataclasses}])
+    check_solver_result(transaction, [{"job": "install", "package": dataclasses}])
 
 
-def test_solver_can_resolve_transitive_extras(solver, repo, package):
+def test_solver_can_resolve_transitive_extras(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(Factory.create_dependency("requests", "^2.24.0"))
     package.add_dependency(Factory.create_dependency("PyOTA", "^2.1.0"))
 
@@ -2712,10 +3322,10 @@ def test_solver_can_resolve_transitive_extras(solver, repo, package):
     repo.add_package(get_package("certifi", "2017.4.17"))
     repo.add_package(get_package("pyopenssl", "0.14"))
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": get_package("certifi", "2017.4.17")},
             {"job": "install", "package": get_package("pyopenssl", "0.14")},
@@ -2725,7 +3335,9 @@ def test_solver_can_resolve_transitive_extras(solver, repo, package):
     )
 
 
-def test_solver_can_resolve_for_packages_with_missing_extras(solver, repo, package):
+def test_solver_can_resolve_for_packages_with_missing_extras(
+    solver: Solver, repo: Repository, package: ProjectPackage
+):
     package.add_dependency(
         Factory.create_dependency(
             "django-anymail", {"version": "^6.0", "extras": ["postmark"]}
@@ -2748,10 +3360,10 @@ def test_solver_can_resolve_for_packages_with_missing_extras(solver, repo, packa
     repo.add_package(boto3)
     repo.add_package(requests)
 
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": django},
             {"job": "install", "package": requests},
@@ -2761,7 +3373,7 @@ def test_solver_can_resolve_for_packages_with_missing_extras(solver, repo, packa
 
 
 def test_solver_can_resolve_python_restricted_package_dependencies(
-    solver, repo, package, locked
+    solver: Solver, repo: Repository, package: Package, locked: Repository
 ):
     package.add_dependency(
         Factory.create_dependency("futures", {"version": "^3.3.0", "python": "~2.7"})
@@ -2782,10 +3394,10 @@ def test_solver_can_resolve_python_restricted_package_dependencies(
     repo.add_package(futures)
     repo.add_package(pre_commit)
 
-    ops = solver.solve(use_latest=["pre-commit"])
+    transaction = solver.solve(use_latest=["pre-commit"])
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": futures},
             {"job": "install", "package": pre_commit},
@@ -2794,7 +3406,7 @@ def test_solver_can_resolve_python_restricted_package_dependencies(
 
 
 def test_solver_should_not_raise_errors_for_irrelevant_transitive_python_constraints(
-    solver, repo, package
+    solver: Solver, repo: Repository, package: Package
 ):
     package.python_versions = "~2.7 || ^3.5"
     solver.provider.set_package_python_versions("~2.7 || ^3.5")
@@ -2831,13 +3443,67 @@ def test_solver_should_not_raise_errors_for_irrelevant_transitive_python_constra
     repo.add_package(pre_commit)
     repo.add_package(importlib_resources)
     repo.add_package(importlib_resources_3_2_1)
-    ops = solver.solve()
+    transaction = solver.solve()
 
     check_solver_result(
-        ops,
+        transaction,
         [
             {"job": "install", "package": importlib_resources_3_2_1},
             {"job": "install", "package": pre_commit},
             {"job": "install", "package": virtualenv},
+        ],
+    )
+
+
+@pytest.mark.parametrize("is_locked", [False, True])
+def test_solver_keeps_multiple_locked_dependencies_for_same_package(
+    solver: Solver,
+    repo: Repository,
+    package: Package,
+    locked: Repository,
+    is_locked: bool,
+):
+    solver.provider.set_package_python_versions("^3.6")
+    package.add_dependency(
+        Factory.create_dependency("A", {"version": "~1.1", "python": "<3.7"})
+    )
+    package.add_dependency(
+        Factory.create_dependency("A", {"version": "~1.2", "python": ">=3.7"})
+    )
+
+    a11 = Package("A", "1.1")
+    a12 = Package("A", "1.2")
+
+    a11.add_dependency(Factory.create_dependency("B", {"version": ">=0.3"}))
+    a12.add_dependency(Factory.create_dependency("B", {"version": ">=0.3"}))
+
+    b03 = Package("B", "0.3")
+    b04 = Package("B", "0.4")
+    b04.python_versions = ">=3.6.2,<4.0.0"
+
+    repo.add_package(a11)
+    repo.add_package(a12)
+    repo.add_package(b03)
+    repo.add_package(b04)
+
+    if is_locked:
+        a11_locked = a11.clone()
+        a11_locked.python_versions = "<3.7"
+        locked.add_package(a11_locked)
+        a12_locked = a12.clone()
+        a12_locked.python_versions = ">=3.7"
+        locked.add_package(a12_locked)
+        locked.add_package(b03.clone())
+        locked.add_package(b04.clone())
+
+    transaction = solver.solve()
+
+    check_solver_result(
+        transaction,
+        [
+            {"job": "install", "package": b03},
+            {"job": "install", "package": b04},
+            {"job": "install", "package": a11},
+            {"job": "install", "package": a12},
         ],
     )
