@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import pytest
 
 from poetry.repositories import Pool
@@ -17,7 +19,7 @@ def test_pool_raises_package_not_found_when_no_package_is_found():
 def test_pool():
     pool = Pool()
 
-    assert 0 == len(pool.repositories)
+    assert len(pool.repositories) == 0
     assert not pool.has_default()
 
 
@@ -25,7 +27,7 @@ def test_pool_with_initial_repositories():
     repo = Repository()
     pool = Pool([repo])
 
-    assert 1 == len(pool.repositories)
+    assert len(pool.repositories) == 1
     assert not pool.has_default()
 
 
